test: 체크 로직 테스트
