+++ README.md
@@ -27,8 +27,6 @@
       - [x] 상품 금액 합산
       - [x] 배송비 10만원 이상이면 0원 적용
       - [x] 총 결제 금액 합산
-    - [ ] 선택 상태(initial state) 로직
-      - [ ] 페이지 진입 시 모든 상품이 선택된 상태인지 판단하는 유틸 함수/리듀서
     - [ ] 장바구니 아이템 로딩 로직
       - [ ] API 호출(mock) → 정상 응답을 받아 state에 잘 넣는지
     - [ ] 장바구니 아이템 제거/수량 변경 로직
@@ -41,7 +39,7 @@
     - [ ] “장바구니에 담은 상품이 없습니다.” 텍스트 표시 & 주문 확인 버튼 비활성화
     - [ ] 장바구니 아이템 UI 렌더링
       - [ ] 로딩된 아이템이 화면에 잘 보이는지
-    - [ ] 페이지 진입 시 모든 체크박스가 체크된 상태
+    - [x] 페이지 진입 시 모든 체크박스가 체크된 상태
     - [ ] 장바구니에서 상품 제거 UI
       - [ ] ‘삭제’ 버튼 클릭 → 해당 아이템이 DOM에서 사라지는지
     - [ ] 수량 변경 UI
